requirements.txt
torch>=2.0.0
transformers>=4.30.0
sacrebleu>=2.3.0
sentencepiece>=0.1.99
gradio>=3.40.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0